changed Abstract3D things to Structs made implemented Instances in some places reworking Uniforms again, so the Data is in a Struct and those structs are used directrly as the Uniforms also made the Data Structs be used for Buffer Data Binding

+++ VoidFactory/Surface2D/Chunk2D_Collection.cs
@@ -350,8 +350,8 @@ namespace VoidFactory.Surface2D
                         {
                             tileIdx = new TileIndex(t);
 
-                            abs.Y = chunk_y + (tileIdx.y * Tile_Size);
-                            abs.C = chunk_c + (tileIdx.c * Tile_Size);
+                            abs.Y = (float)(chunk_y + (tileIdx.y * Tile_Size));
+                            abs.C = (float)(chunk_c + (tileIdx.c * Tile_Size));
 
                             dist = (mid - abs).Len;
                             if (dist <= rad)
